+++ libs/ymir-core/src/ymir/hw/vdp/renderer/vdp_renderer_sw.cpp
@@ -2498,6 +2498,10 @@ FORCE_INLINE void SoftwareVDPRenderer::VDP2CalcAccessPatterns(VDP2Regs &regs2) {
                 //                     B   PN3 ...3....          Cyberbots - Fullmetal Madness, in-game
                 //                     A   CP3 ...3....  no      Cyberbots - Fullmetal Madness, in-game
                 //                     B   CP3 .......7  no      Cyberbots - Fullmetal Madness, in-game
+                // 17 lo   1x  pal256  B1  PN1 ..2.45..          BattleSport, loading screen
+                //                     B1  CP1 ......67  no      BattleSport, loading screen
+                // 18 lo   1x  pal256  B1  PN3 ..2.45..          Daisuki, intro animation
+                //                     B1  CP3 ......67  no      Daisuki, intro animation
                 // clang-format on
                 //
                 // Seems like the bitmap "delay" is caused by configuring out-of-phase reads for an NBG in different
@@ -2518,6 +2522,9 @@ FORCE_INLINE void SoftwareVDPRenderer::VDP2CalcAccessPatterns(VDP2Regs &regs2) {
                 // In case #15, the CP2 access in bank A0 is assigned to T3, which is illegal for PN at T0.
                 // Case #16 shows legal accesses. Note that there are CP0-CP3 accesses in both the T0-T3 and T4-T7
                 // ranges, but this does not cause the T4-T7 accesses to be shifted.
+                // Cases #17 and #18 have more PN accesses than necessary and show that only the first PN access matters
+                // for the delay checks. In both cases, the first PN access occurs on T2, which makes the CP accesses in
+                // T6 and T7 valid. PN accesses on T4 and T5 would make those CP accesses invalid.
 
                 auto &bgParams = regs2.bgParams[bgIndex + 1];
                 if (bgParams.bitmap && hires) {
@@ -2567,9 +2574,7 @@ FORCE_INLINE void SoftwareVDPRenderer::VDP2CalcAccessPatterns(VDP2Regs &regs2) {
         }
 
         // Apply the delay
-        if (bgPN == 0) {
-            bgParams.charPatDelay = true;
-        } else if (hires) {
+        if (hires) {
             // Valid character pattern access masks per timing for high resolution modes
             static constexpr uint8 kPatterns[2][4] = {
                 // 1x1 character patterns
@@ -2585,9 +2590,10 @@ FORCE_INLINE void SoftwareVDPRenderer::VDP2CalcAccessPatterns(VDP2Regs &regs2) {
                 // Delay happens when either:
                 // - CP access happens entirely before PN access
                 // - CP access occurs in illegal time slot
-                if ((bgPN & (1u << pnIndex)) != 0 &&
-                    (bgCP < bgPN || (bgCP & kPatterns[bgParams.cellSizeShift][pnIndex]) != bgCP)) {
-                    bgParams.charPatDelay = true;
+                if ((bgPN & (1u << pnIndex)) != 0) {
+                    if (bgCP < bgPN || (bgCP & kPatterns[bgParams.cellSizeShift][pnIndex]) != bgCP) {
+                        bgParams.charPatDelay = true;
+                    }
                     break;
                 }
             }
@@ -2608,8 +2614,8 @@ FORCE_INLINE void SoftwareVDPRenderer::VDP2CalcAccessPatterns(VDP2Regs &regs2) {
                     }
                     if ((bgCP & kPatterns[pnIndex]) == 0) {
                         bgParams.charPatDelay = true;
-                        break;
                     }
+                    break;
                 }
             }
         }